Ensure name in env var matches name in args
This prevents someone from making an error like:

DISTRIBUTOR_FOO=foo,3000
DISTRIBUTOR_FOO2=foo,3001

That would cause two listeners on the same server name with nothing to differentiate them.

Instead, the user must use:

DISTRIBUTOR_FOO=foo,3000
DISTRIBUTOR_FOO2=foo2,3001

And ensure that traffic intended for 3001 comes in on the domain foo2

We don't simply pull the name from the suffix in order to allow users to have:

DISTRIBUTOR_BAR=Bar,3000

Which is gross but I dunno maybe it's necessary for some reason. Seems overly proscriptive to forbid it.